requirements.txt

pandas==2.1.3
streamlit==1.28.2
spire-presentation==9.1.0
python-pptx==0.6.23



